+++ meteor/client/ui/Settings/ShowStyle/VariantSettings.tsx
@@ -9,7 +9,7 @@ import { MeteorCall } from '../../../../lib/api/methods'
 import { ShowStyleBase } from '../../../../lib/collections/ShowStyleBases'
 import { ShowStyleVariant } from '../../../../lib/collections/ShowStyleVariants'
 import { doModalDialog } from '../../../lib/ModalDialog'
-import { SourceLayerDropdownOption } from '../BlueprintConfigManifest'
+import { SourceLayerDropdownOption } from '../BlueprintConfigSchema'
 import { ShowStyleVariantId } from '@sofie-automation/corelib/dist/dataModel/Ids'
 import { NoticeLevel, NotificationCenter, Notification } from '../../../../lib/notifications/notifications'
 import { UploadButton } from '../../../lib/uploadButton'
