+++ meteor/client/ui/Settings/ShowStyleBaseSettings.tsx
@@ -7,7 +7,7 @@ import { ShowStyleVariant } from '../../../lib/collections/ShowStyleVariants'
 import RundownLayoutEditor from './RundownLayoutEditor'
 import { Studio, MappingsExt } from '../../../lib/collections/Studios'
 import { BlueprintManifestType, IShowStyleConfigPreset, ISourceLayer } from '@sofie-automation/blueprints-integration'
-import { SourceLayerDropdownOption } from './BlueprintConfigManifest'
+import { SourceLayerDropdownOption } from './BlueprintConfigSchema'
 import { RundownLayoutsAPI } from '../../../lib/api/rundownLayouts'
 import { TriggeredActionsEditor } from './components/triggeredActions/TriggeredActionsEditor'
 import { SourceLayerSettings } from './ShowStyle/SourceLayer'
